Fix more MSVC warnings.

+++ test/c/save_load.c
@@ -34,15 +34,15 @@ void save_load_load_equals_save(CuTest *tc)
 	CuAssertTrue(tc, save.pImpl != load.pImpl);
 
 	/* Check degree. */
-	CuAssertIntEquals(tc, ts_bspline_degree(&save),
+	CuAssertIntEquals(tc, (int) ts_bspline_degree(&save),
 		(int) ts_bspline_degree(&load));
 
 	/* Check dimension. */
-	CuAssertIntEquals(tc, ts_bspline_dimension(&save),
+	CuAssertIntEquals(tc, (int) ts_bspline_dimension(&save),
 		(int) ts_bspline_dimension(&load));
 
 	/* Check knots. */
-	CuAssertIntEquals(tc, ts_bspline_num_knots(&save),
+	CuAssertIntEquals(tc, (int) ts_bspline_num_knots(&save),
 		(int) ts_bspline_num_knots(&load));
 	for (i = 0; i < ts_bspline_num_knots(&save); i++) {
 		C(ts_bspline_knot_at(&save, i, &knot[0], &status))
@@ -51,7 +51,7 @@ void save_load_load_equals_save(CuTest *tc)
 	}
 
 	/* Check control points. */
-	CuAssertIntEquals(tc, ts_bspline_num_control_points(&save),
+	CuAssertIntEquals(tc, (int) ts_bspline_num_control_points(&save),
 		(int) ts_bspline_num_control_points(&load));
 	CuAssertIntEquals(tc, 2, (int) ts_bspline_dimension(&save));
 	for (i = 0; i < ts_bspline_num_control_points(&save); i++) {
